+++ GuidancerUI/CollectionViewCell/CollectionViewCell.swift
@@ -12,10 +12,12 @@ class CollectionViewCell: UICollectionViewCell {
     
     //MARK: - var/let
     let collectionCell = UICollectionViewCell()
-    let padding: CGFloat = 50.0
+    let padding: CGFloat = 50.0 //TODO: - clear
     let spacing: CGFloat = 25.0
     
-    
+    func setupContent() {
+        secondSV.setupContent(whoAndWhenPost: <#T##String#>)
+    }
     
     //MARK: - Init
     
@@ -34,7 +36,7 @@ class CollectionViewCell: UICollectionViewCell {
     
     
     
-        let firstSV: verticalSV = {
+        let firstSV: verticalSV = { //TODO: - hide in class
             let sv = verticalSV()
             sv.spacing = CGFloat(5.0)
             sv.distribution = .equalCentering
@@ -44,8 +46,8 @@ class CollectionViewCell: UICollectionViewCell {
             return sv
         }()
 
-        let secondSV: horizontalSV = {
-            let sv = horizontalSV()
+        private let secondSV: HorizontalSV = {
+            let sv = HorizontalSV()
             sv.spacing = CGFloat(5.0)
             sv.distribution = .fill
             sv.alignment = .center
@@ -89,6 +91,62 @@ class CollectionViewCell: UICollectionViewCell {
     
 }
 
+
+fileprivate class HorizontalSV: UIStackView { //TODO: - change class name
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setup()
+    }
+    
+    required init(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func setupContent(whoAndWhenPost: String) { //TODO: - setupContent
+        self.whoAndWhenPost.text = whoAndWhenPost
+    }
+    
+    let whoAndWhenPost: GLabel = {
+        let label = GLabel(text: "Feliks Omarov 2 дня назад",
+                           font: UIFont.systemFont(ofSize: 12, weight: .light))
+        label.textColor = UIColor.gGray
+        return label
+    }()
+
+    let likesNumber: GLabel = {
+        let label = GLabel(text: "228",
+                           font: .regular14)
+        label.textColor = .gBlack
+        return label
+    }()
+    
+    let commentNumber: GLabel = {  // TODO: - Refactoer
+        let label = GLabel(text: "228",
+                           font: UIFont.systemFont(ofSize: 12, weight: .regular))
+        label.textColor = .gBlack
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    
+    let heartImage = GLikeCommentButton(icon: <#T##UIImage?#>)
+    
+    let commentImage: GLikeCommentButton = { //TODO: - refactore
+//        let image = GLikeCommentButton(systemNameImage: "text.bubble")
+        image.tintColor = .black
+        image.translatesAutoresizingMaskIntoConstraints = false
+        return image
+    }()
+    
+    func setup() {
+        addArrangedSubview(whoAndWhenPost)
+        addArrangedSubview(likesNumber)
+        addArrangedSubview(heartImage)
+        addArrangedSubview(commentNumber)
+        addArrangedSubview(commentImage)
+    }
+}
