+++ GuidancerUI/CollectionViewCell/HorizontalSV.swift
@@ -1,0 +1,9 @@
+//
+//  horizontalSV.swift
+//  Guidancer
+//
+//  Created by Артём Коротков on 03.11.2022.
+//
+
+import UIKit
+
